+++ api/server.js
@@ -25,7 +25,7 @@ server.use("/api/auth", AuthRouter);
 server.use("/api/users", passport.authenticate("jwt", { session: false }), UserRouter);
 server.use("/api/teams", passport.authenticate("jwt", { session: false }), TeamRouter);
 server.use("/api/videos", passport.authenticate("jwt", { session: false }), VideoRouter);
-server.use("/api/invites", passport.authenticate("jwt", { session: false }), InviteRouter);
+server.use("/api/invites", InviteRouter);
 server.use("/api/avatars", AvatarRouter);
 
 server.use("/public", express.static(path.join(__dirname, "../public")));
